weight NULL allowance
Former-commit-id: 11380434046cc0459f9fae1b5ced63a9e1154907
Former-commit-id: a2a400aee8c0e64f3938e89c735124c2a2c18b80

+++ src/cpu/matrix/matrix_dense.cpp
@@ -153,7 +153,7 @@ MatrixDense<T>::MatrixDense(int sharedA, int me, int wDev, char ord, size_t m, s
 
   ASSERT(ord == 'r' || ord == 'R' || ord == 'c' || ord == 'C');
   _ord = (ord == 'r' || ord == 'R') ? ROW : COL;
-  DEBUG_FPRINTF(stderr,"ord=%c m=%zu n=%zu mValid=%zu\n",ord,(int)m,(int)n,int(mValid));
+  DEBUG_FPRINTF(stderr,"ord=%c m=%zu n=%zu mValid=%zu\n",ord,m,n,mValid);
 
   CpuData<T> *info = new CpuData<T>(data); // new structure (holds pointer to data and GPU handle)
   CpuData<T> *infoy = new CpuData<T>(datay); // new structure (holds pointer to data and GPU handle)
